move addMultipleFeatures fn to utils
